Fully matching BootLayout, work on SceneStateSelectMode
also moved some stuff around and added In2DAreaMoveControl
