+++ lib/game/layoutActors/CoinCounter/CoinCounter.h
@@ -1,0 +1,9 @@
+#pragma once
+
+/*
+ * VTable Loc: 1CC3170
+ */
+
+#include "al/layout/LayoutActor.h"
+
+class CoinCounter : public al::LayoutActor {};
