--- lib/game/layouts/CoinCounter.h
@@ -1,9 +1,0 @@
-#pragma once
-
-/*
- * VTable Loc: 1CC3170
- */
-
-#include "al/layout/LayoutActor.h"
-
-class CoinCounter : public al::LayoutActor {};
