+++ lib/game/scenes/FirstSequenceScene/FirstSequenceScene.h
@@ -1,5 +1,15 @@
 #pragma once
 
-class FirstSequenceScene : public al::Scene {
+#include "al/scene/Scene.h"
 
+class FirstSequenceScene : public al::Scene {
+    FirstSequenceScene(const char* name);
+    ~FirstSequenceScene();
+
+    void init(const al::SceneInitInfo&);
+    void appear() override;
+    void control() override;
+    void drawMain() override;
+
+    void setNoWipe();
 };
