+++ src/game/actor/dimension/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(odyssey PRIVATE
+    ActorDimensionKeeper.cpp
+    In2DAreaMoveControl.cpp
+)
