+++ src/game/actor/dimension/In2DAreaMoveControl.cpp
@@ -1,0 +1,5 @@
+#include "game/actor/dimension/In2DAreaMoveControl.h"
+
+In2DAreaMoveControl::In2DAreaMoveControl() {
+
+}
