+++ src/game/layoutActors/BootLayout/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(odyssey PRIVATE
+    BootLayout.cpp
+)
