+++ src/game/layoutActors/CMakeLists.txt
@@ -1,3 +1,1 @@
-target_sources(odyssey PRIVATE
-    MapLayout.cpp
-)
+add_subdirectory(BootLayout)
