--- src/game/layoutActors/MapLayout.cpp
@@ -1,1 +1,0 @@
-#include "game/Layouts/MapLayout.h"
