+++ src/game/liveActors/Mofumofu/MofumofuWarpHole.cpp
@@ -1,4 +1,4 @@
-#include "game/objects/Mofumofu/MofumofuWarpHole.h"
+#include "game/liveActors/Mofumofu/MofumofuWarpHole.h"
 
 #include <math/seadQuat.h>
 #include <math/seadVector.h>
