+++ src/game/scenes/FirstSequenceScene/CMakeLists.txt
@@ -1,3 +1,4 @@
 target_sources(odyssey PRIVATE
     FirstSequenceScene.cpp
+    SceneStateSelectMode.cpp
 )
