+++ src/game/scenes/FirstSequenceScene/SceneStateSelectMode.cpp
@@ -1,0 +1,1 @@
+#include "game/scenes/FirstSequenceScene/SceneStateSelectMode.h"
